requirements.txt
# Core Dependencies
python-binance==1.0.19
pandas==2.1.4
numpy==1.26.2


# Environment Management
python-dotenv==1.0.0

# Data Visualization (for backtesting)
matplotlib==3.8.2
seaborn==0.13.0

# Logging and Monitoring
requests==2.31.0

# Optional: For Telegram notifications
python-telegram-bot==20.7

# Testing
pytest==7.4.3

python-dotenv